rendering.common is renamed to rendering.context
