+++ pyviews/containers.py
@@ -5,7 +5,7 @@ from pyviews.core.binding import Bindable
 from pyviews.core.rendering import Node, NodeGlobals, RenderingContext
 from pyviews.core.xml import XmlNode
 from pyviews.pipes import apply_attributes, render_children
-from pyviews.rendering.common import get_child_context
+from pyviews.rendering.context import get_child_context
 from pyviews.rendering.pipeline import RenderingPipeline, render, render_view
