+++ pyviews/presenter.py
@@ -6,7 +6,7 @@ from typing import Dict, Optional, Union
 from pyviews.core.rendering import InstanceNode, Node, NodeGlobals, RenderingContext
 from pyviews.core.xml import XmlNode
 from pyviews.pipes import apply_attributes, render_children
-from pyviews.rendering.common import get_child_context
+from pyviews.rendering.context import get_child_context
 from pyviews.rendering.pipeline import RenderingPipeline
